Se agrega seccion de reportes
Se crean reportes para ser expotados excel y pdf

+++ VentasProyect/Models/Ventas/Ventas.cs
@@ -14,7 +14,7 @@ namespace VentasProyect.Models.Ventas
         public string usu_nombre { get; set; }
         public Nullable<System.DateTime> ven_fecha { get; set; }
         public string ven_metodo_pago { get; set; }
-        public Nullable<long> ven_total { get; set; }
-        public Nullable<long> ven_numero_transaccion { get; set; }
+        public Nullable<int> ven_total { get; set; }
+        public Nullable<int> ven_numero_transaccion { get; set; }
     }
 }
